ajout du projet front

+++ src/main/java/com/beezy/websoa/services/EtudiantServices.java
@@ -8,10 +8,10 @@ import com.beezy.websoa.entities.Unite;
 public interface EtudiantServices {
 	
 	public List<Etudiant> listEtudiant();
-	public List<Unite> listUniteByIdEtudiant(int id);
-	public void delete(int id);
+	public List<Unite> listUniteByIdEtudiant(Long id);
+	public void delete(Long id);
 	public boolean saveOrUpdate(Etudiant etudiant);
-	public Etudiant getEtudiantById(int id);
+	public Etudiant getEtudiantById(Long id);
 	public void addEtudiant(Etudiant e);
 	public void updateEtudiant(Etudiant e);
